Update to 2.3.10

+++ mewwalletbl/src/main/java/com/myetherwallet/mewwalletbl/data/database/EntitySwap.kt
@@ -6,6 +6,7 @@ import com.myetherwallet.mewwalletbl.core.persist.database.dao.ExchangeDao
 import com.myetherwallet.mewwalletbl.data.Blockchain
 import com.myetherwallet.mewwalletbl.data.api.TransactionStatus
 import com.myetherwallet.mewwalletkit.bip.bip44.Address
+import java.math.BigDecimal
 import java.util.*
 
 @Entity(tableName = ExchangeDao.TABLE_NAME)
@@ -15,8 +16,8 @@ data class EntitySwap(
     val updateTime: Date,
     val fromDescriptionId: Long,
     val toDescriptionId: Long,
-    val fromAmount: Double,
-    val toAmount: Double,
+    val fromAmount: BigDecimal,
+    val toAmount: BigDecimal,
     val accountId: Long,
     val toAddress: Address,
     val dex: String,
